grid.text() handles any language object (not just expressions)
extension arg for dataViewport() can be vector 

justifications can be numeric 


git-svn-id: https://svn.r-project.org/R/trunk@32019 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/library/grid/src/viewport.c
@@ -68,12 +68,12 @@ SEXP viewportLayout(SEXP vp) {
     return VECTOR_ELT(vp, VP_LAYOUT);
 }
 
-int viewportHJust(SEXP vp) {
-    return INTEGER(VECTOR_ELT(vp, VP_VALIDJUST))[0];
+double viewportHJust(SEXP vp) {
+    return REAL(VECTOR_ELT(vp, VP_VALIDJUST))[0];
 }
 
-int viewportVJust(SEXP vp) {
-    return INTEGER(VECTOR_ELT(vp, VP_VALIDJUST))[1];
+double viewportVJust(SEXP vp) {
+    return REAL(VECTOR_ELT(vp, VP_VALIDJUST))[1];
 }
 
 SEXP viewportLayoutPosRow(SEXP vp) {
